common_tag/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.15.0
chardet==3.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
Django==3.1.4
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.12.2
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
fernet==1.0.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.3.6
MarkupSafe==1.1.1
mysqlclient==2.1.0
numpy==1.22.2
packaging==20.8
Pillow==9.0.1
pyaes==1.6.1
pycparser==2.21
PyJWT==2.3.0
pyparsing==2.4.7
pytz==2020.4
qrcode==7.3.1
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
